pricing_engine/requirements.txt
Flask==3.0.3
pymysql==1.1.1
python-dotenv==1.0.1
pandas==2.2.2
numpy==2.0.2
statsmodels==0.14.2
lightgbm==4.5.0
scikit-learn==1.5.2
joblib==1.4.2
pytest==8.3.3
